[refactor] util 폴더 생성
