+++ app/src/main/java/org/sopt/dosopttemplate/presentation/util/ItemDifCallback.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data
+package org.sopt.dosopttemplate.presentation.util
 
 import androidx.recyclerview.widget.DiffUtil
